Add files via upload
Collaborative filtering is a method of making automatic predictions about the interests of a user by collecting preferences from many users. Big data is a large volume of structured, semi-structured and unstructured data that has the potential to be mined for information.By using clustering with collaborative filtering in big data application like online shopping, the execution time is reduced.The users will be provided with the search results and some additional results that help them to save time.